fix navbar for async
